New Development: No             JIRA Issue: CAS-3517,4621         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: New calibrater_cmpt methods, mainly          Test Programs: tCLPatchPanel   Put in Release Notes: Yes              Module(s): synthesis, calibrater tool            Description: Initial check-in of cal library c++, mainly to 			ensure happy co-existence with existing 			calibration apply framework.  Additional 			callib deployment (in tasks, etc., w/ 			unit tests and regresions) coming soon.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/code/synthesis@25991 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ CalTables/NewCalTable.cc
@@ -527,19 +527,19 @@ void NewCalTable::writeToDisk(const String& outTableName)
   sorted.deepCopy(outTableName,Table::New);
 };
 
-Complex NewCalTable::NCTtestvalueC(Int iant,Int ispw,Int ich,Double time,Double refTime,Double tint) {
+Complex NewCalTable::NCTtestvalueC(Int iant,Int ispw,Double ich,Double time,Double refTime,Double tint) {
 
-  Double a=1.0 + Double(iant)/10.0 + Double(ispw)/100.0 +Double(ich)/10000.0;
+  Double a=1.0 + Double(iant)/10.0 + Double(ispw)/100.0 + ich/10000.0;
   Double dt=(time-refTime)/tint;
-  Double p=dt+Double(ich)/100.0; 
+  Double p=dt+ich/100.0; 
   return Complex(Float(a*cos(p)),Float(a*sin(p)));
 
 }
 
-Float NewCalTable::NCTtestvalueF(Int iant,Int ispw,Int ich,Double time,Double refTime,Double tint) {
+Float NewCalTable::NCTtestvalueF(Int iant,Int ispw,Double ich,Double time,Double refTime,Double tint) {
 
   Double dt=(time-refTime)/tint;
-  Double a=dt + Double(iant)/10.0 + Double(ispw)/100.0 +Double(ich)/10000.0;
+  Double a=dt + Double(iant)/10.0 + Double(ispw)/100.0 + ich/10000.0;
   return Float(a);
 
 }
